refactor(prefer-const): extract logic into class
After having finished the implementation, I've decided to extract the
logic into a function, since I strongly believe it'll be easier to
understand, as functions are equivalent to classes, if done right.
The reason of this commit is to save the current implementation in case
I want to revert changes.