test(prefer-const): add original rule tests & fix code
By adding the original rule tests, I was able to identify issues with
the rule, that had been caused because of the refactoring. Now, not only
I think the code is more readable and maintainable, but also the
behavior is as expected, and there are tests to ensure that.

There's a tests that has been skipped since it reporting a false
positive, which is dues to `@typescript-eslint`. I've reported the issue
in https://github.com/typescript-eslint/typescript-eslint/issues/10426